Warn about out-of-bounds accesses for null-terminated array pointers. (#415)
It is easy to write checked code that tries to write to the null terminator at the end of a null-terminated array, or tries to read elements beyond the null terminator.   These operations will trigger runtime bounds failures.   By default null-terminated array pointers have a bounds of count(0) and we haven't implemented a flow analysis that widens the bounds of these pointers yet..

This adds an analysis that that leverages the existing bounds declaration checking support for constant-sized ranges to detect out-of-bounds errors for these cases.   I refactored some of the existing code so that we can re-use it.   I also added support for determining when two constant-sized ranges partially overlap, or when they are empty, so that we can produce more meaningful error messages.  This addresses issue #407.

Testing:
- Passes testing on x64 Windows locally.
- Passes automated testing on Linux.
- Added new tests for exercising this functionality.
- For existing tests in the Checked C repo which intentionally cause out-of-bounds errors at runtime, this detects some of those cases at compile time and warns about them.  Suppress the warnings (using -Wno-check-memory-access).   Also update some null-pointer tests with expected warnings.  This was covered by a separate commit to the Checked C repo.